requirements/base.txt
# roll_engine requirements
celery>=3.0,<4.0
django-fsm


# rest_client requirements
pyyaml
requests>=2.6.0
elasticsearch

# tars requirements
decisionTable
Django>=1.7,<1.8
django-constance>=1.2
django-picklefield
django-filter <=0.11.0
djangorestframework>=3.0.4,<3.1.0
djangorestframework-camel-case
djangorestframework-jsonp
MySQL-python
sqlparse
redis
raven
kafka-python
